refactor: simplify security setup by removing unused features and consolidating imports
